added a header with styling
Used bootstraps grid system to lay out image and header.
Seperate div that held a group of h1's was creted to prevent using the same piece of code on each h1.
Header font-size changes at viewport max-width 1200px. Also used Prettier to clean up CSS and HTML formatting. A meta tag describing the author and the description of the website was added to the HTML.